+++ resources/scale_common/ansible_compute_playbook/ansible_compute_playbook.tf
@@ -25,6 +25,7 @@ variable "compute_instances_by_id" {}
 variable "compute_filesystem_mountpoint" {}
 variable "compute_cluster_gui_username" {}
 variable "compute_cluster_gui_password" {}
+variable "host" {}
 
 locals {
   tf_inv_path                     = format("%s/%s", "/tmp/.schematics/IBM", "compute_tf_inventory.json")
@@ -113,7 +114,7 @@ resource "null_resource" "call_scale_install_playbook" {
   connection {
     bastion_host = var.bastion_public_ip
     user         = local.bastion_user
-    host         = "0.0.0.0"
+    host         = var.host
     private_key  = file(var.bastion_ssh_private_key)
   }