+++ scripts/user_data_symphony.sh
@@ -34,6 +34,7 @@ export securityGroupID=${securityGroupID}
 export sshkey_ID=${sshkey_ID}
 export regionName=${regionName}
 export zoneName=${zoneName}
+export resourceGroupID=${resourceGroupID}
 #prefix should be 10 characters or fewer
 export hostPrefix=${hostPrefix}
 
@@ -446,7 +447,6 @@ su ${CLUSTERADMIN} -c 'egoconfig join \${PRIMARY_MASTER} -f'
 su ${CLUSTERADMIN} -c 'egoconfig addresourceattr "[resourcemap ibmcloud*cloudprovider] [resource corehoursaudit]"'
 echo "source ${EGO_TOP}/profile.platform" >> /root/.bashrc
 sleep $STARTUP_DELAY
-chmod 0755 /usr/bin/pkexec
 systemctl start ego
 echo END >> /var/log/postprovisionscripts.log 2>&1
 EOF
@@ -468,6 +468,7 @@ cat << EOF > $IBM_CLOUD_TEMPLATE_FILE
     "imageId": "${imageID}",
     "subnetId": "${subnetID}",
     "vpcId": "${vpcID}",
+    "resourceGroupId": "${resourceGroupID}",
     "vmType": "${hf_profile}",
     "securityGroupIds": ["${securityGroupID}"],
     "sshkey_id": "${sshkey_ID}",
@@ -723,11 +724,6 @@ function start_ego
     systemctl start ego
 }
 
-function polkit
-{
-  # This changes are made to check the security vulnerability
-  chmod 0755 /usr/bin/pkexec
-}
 ##################################################################
 
 if [ -z "${egoHostRole}" ]; then
@@ -758,7 +754,6 @@ if [ "${egoHostRole}" == "primary" ]; then
     scale_append_hosts_file
     update_passwords
     wait_for_candidate_hosts
-    polkit
     rm -f $DONE_FILE
 elif [ "${egoHostRole}" == "secondary" ]; then
     mount_nfs
@@ -773,7 +768,6 @@ elif [ "${egoHostRole}" == "secondary" ]; then
     start_ego
     wait_for_management_hosts
     scale_append_hosts_file
-    polkit
 elif [ "${egoHostRole}" == "management" ]; then
     mount_nfs
     wait_for_nfs
@@ -787,7 +781,6 @@ elif [ "${egoHostRole}" == "management" ]; then
     start_ego
     wait_for_management_hosts
     scale_append_hosts_file
-    polkit
 elif [ "${egoHostRole}" == "scale_storage" ]; then
     config_hyperthreading
     mount_nfs
@@ -799,7 +792,6 @@ elif [ "${egoHostRole}" == "scale_storage" ]; then
     copy_sshkey
     wait_for_management_hosts
     scale_append_hosts_file
-    polkit
 else
     config_hyperthreading
     mount_nfs_readonly
@@ -816,5 +808,4 @@ else
     start_ego
     wait_for_management_hosts
     scale_append_hosts_file
-    polkit
 fi