+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # **CHANGELOG**
 
+## **1.2.2**
+### **CHANGES**
+- RHEL 8.4 Custom image updated with polkit vulnerability fix.
+- Fixed bug to use users custom image for scale storage nodes. 
+- Fixed bug to create dynamic host with different resource group.
+- Fixed fip issue with ssh allowed ips provided.
+- Scale version has been updated from 5.1.2 to 5.1.3.1
+- Terraform version has been updated from v0.14 to v1.1
+- Wait duration time for storage and compute set up has been increased to 180s
+
 ## **1.2.1**
 ### **CHANGES**
 - Changes to post provisioning scripts to mitigate Polkit Local Privilege Escalation Vulnerability (CVE-2021-4034).